requirements.txt
Flask
gunicorn
Pillow
urllib3